Remove unused --split-product-types flag (close #105)
Currently we have code in place to detect a `--split-product-types`
flag, and pass it along to the code behind `igluctl static generate`:

https://docs.snowplowanalytics.com/docs/pipeline-components-and-applications/iglu/igluctl-2/#static-generate

However, the code to parse command line options doesn't currently
look for `--split-product-types`, but rather for `--raw-mode`.
Thanks to this PR for bringing this to our attention:

https://github.com/snowplow-incubator/igluctl/pull/99/files

We pass along a split product val into the code to generate DDL
files, but don't end up using it.

From the point of view of a user running `igluctl static generate`,
currently passing the `--split-product-types` flag returns an error.

With this change, passing the `--split-product-types` flag still
returns an error, but we get to remove code that isn't currently
in use.