Option to filter schemas when linting (close #86) (#111)
There are times when a directory contains schemas with known issues,
but it is still helpful to run lint on the other schemas with full
strictness. So, here we introduce the functionality to be able to run e.g.

`igluctl lint --skip-schemas iglu:com.acme/click/jsonschema/1-0-1,iglu:com.acme/scroll/jsonschema/1-0-1`

Which would lint all the schemas except the ones passed comma separated
with `--skip-schemas`.