Drop doobie-postgres (close #122)
It turns out that we are pulling in both `doobie-core` and `doobie-postgres` into
the build, but if we look at what we are actually using in the code, the only
doobie imports are in `Storage.scala` and they are both from `doobie-core` so
therefore currently we are pulling in `doobie-postgres` and not using it.

This means we were not in practice impacted by

https://security.snyk.io/vuln/SNYK-JAVA-ORGPOSTGRESQL-2390459

in igluctl, but we can tidy up and remove Snyk flagging it as such.